Move files into src.
