--- index.ts
@@ -1,12 +1,0 @@
-export * from "./client.js";
-export * from "./api-client.js";
-export * from "./org.js";
-export * from "./catalog.js";
-export * from "./cortex.js";
-export * from "./chat.js";
-export * from "./content.js";
-export * from "./document.js";
-export * from "./indexers/directory-indexer.js";
-export * from "./indexers/json-indexer.js";
-export * from "./indexers/tsv-indexer.js";
-export * from "./indexers/shopify-indexer.js";
