+++ src/indexers/directory-indexer.ts
