+++ src/test_data/test_large_docx_file.docx
