+++ vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    exclude: ["**/*.js"],
-    setupFiles: ["./test-setup.ts"],
+    include: ["**/*.test.ts"],
+    setupFiles: ["./src/test-setup.ts"],
   },
 });
